pyproject.toml
[tool.poetry]
name = "llm-rag-sql-m5"
version = "0.1.0"
description = "Used as a test to ask questions about the m5 forecasting data in a postgres model"
authors = ["Harmen <harmen.kerkenaar@qbis.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.1.1"
python-dotenv = "^1.0.0"
langchain-openai = "^0.0.2.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
